Refactor and enhance components for better data handling
This commit refactors several components to improve data handling and flexibility. It primarily modifies the way data is passed as props to the components. By taking advantage of the flexibility that comes with the use of props, components such as HeroSection, FeaturesStatsAlt, and TestimonialsSectionAlt have been enhanced to adjust and render content dynamically. This creates an improvement in code reusability and maintainability across the application. Additionally, optional rendering has been added for secondary elements such as subtitles and buttons.